% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc_coding_potential}
\alias{calc_coding_potential}
\title{calc_coding_potential}
\usage{
calc_coding_potential(transcript)
}
\arguments{
\item{transcript}{the introns file}

\item{start_codon}{the start codon location}

\item{end_codon}{the stop codon location}
}
\value{
the LGG socre
}
\description{
calc_coding_potential
}
