% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_tx}
\alias{check_tx}
\title{checking if the transcript without the UTR has a start codon in the first codon and a stop codon in the last codon}
\usage{
check_tx(seq)
}
\arguments{
\item{seq}{The DNA sequence in character form to analyze}
}
\value{
TRUE or FALSE dependent upon whether the transcript has start codon at first codon and stop codon at last codon
}
\description{
checking if the transcript without the UTR has a start codon in the first codon and a stop codon in the last codon
}
