% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{choose_exons}
\alias{choose_exons}
\title{choose_exons}
\usage{
choose_exons(target_junc, exons_by_gene, gene_tar)
}
\arguments{
\item{target_junc}{the target junction to use for analysis}

\item{exons_by_gene}{the exons per gene in the genome}

\item{gene_tar}{a vector pair of gene targets, left is first or only gene, right is second gene. Single element vector if only one gene.}
}
\value{
the unique start and end junction exons
}
\description{
choose_exons
}
