% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{choose_transcripts}
\alias{choose_transcripts}
\title{choosing the transcripts that the exons can exist in}
\usage{
choose_transcripts(exons, transcripts_1, transcripts_2 = c(), exons_by_tx)
}
\arguments{
\item{exons}{the exons for the start and end of the junction}

\item{transcripts_1}{the character vector of transcripts for the start gene}

\item{transcripts_2}{the character vector of transcripts for the end gene}

\item{exons_by_tx}{the exons per transcript}
}
\value{
a list of start junction and end junction transcripts that can be paired
}
\description{
choosing the transcripts that the exons can exist in
}
