% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{countKmers}
\alias{countKmers}
\title{countKmers}
\usage{
countKmers(seqs_vec, K)
}
\arguments{
\item{seqs_vec}{vector of sequences (i.e. data_canon[,2])}

\item{K}{the kmer length}
}
\value{
a list of integer vectors with kmer character vectors as names
}
\description{
countKmers
}
