% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_cds}
\alias{create_cds}
\title{creating the combination cds for the junction}
\usage{
create_cds(combo_exons, UTR5, UTR3, tx_junc)
}
\arguments{
\item{combo_exons}{the combined exons for the two transcripts associated with the junction}

\item{UTR5}{the 5' UTR for the first transcript}

\item{UTR3}{the 3' UTR for the final transcript}

\item{target_junc}{the target junction character vector in form c(chr,strand,start,end)}
}
\description{
creating the combination cds for the junction
}
