% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_orfs}
\alias{find_orfs}
\title{find_orfs}
\usage{
find_orfs(sequ, tx_junc_loc)
}
\arguments{
\item{sequ}{The DNA sequence in character form to analyze}

\item{tx_junc_loc}{The junction location in the transcript}
}
\value{
a vector of error type, protein if translatable or DNA sequence, ensembl transcript name , junction location relative to peptide start, modified transcript length, start and stop codon locations
}
\description{
find_orfs
}
