% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{format_introns}
\alias{format_introns}
\title{format_introns}
\usage{
format_introns(introns)
}
\arguments{
\item{introns}{the introns file}
}
\value{
splicemutr formatted introns file
}
\description{
format_introns
}
