% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getUniqKmers}
\alias{getUniqKmers}
\title{returns the unique set of kmers from a vector of non-unique kmers with
integer counts per junction-tx pair}
\usage{
getUniqKmers(kmers)
}
\arguments{
\item{kmers}{character vector of non-unique kmers}
}
\value{
vector of character vectors representing unique list of kmers from the union of all sequences
}
\description{
returns the unique set of kmers from a vector of non-unique kmers with
integer counts per junction-tx pair
}
