% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is_UTR}
\alias{is_UTR}
\title{is_UTR}
\usage{
is_UTR(target_junc, UTR5, UTR3, junc)
}
\arguments{
\item{target_junc}{the target junction in form chr,strand,start,end}

\item{UTR5}{the 5' UTR}

\item{UTR3}{the 3' UTR}

\item{junc}{the original junc call}
}
\value{
two logicals indicating whether the junction is UTR or not
}
\description{
is_UTR
}
