% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{junc_loc_pep}
\alias{junc_loc_pep}
\title{junc_loc_pep}
\usage{
junc_loc_pep(tx_junc_loc, tx, start_loc, stop_loc)
}
\arguments{
\item{tx_junc_loc}{The junction location in the transcript}

\item{tx}{The transcript sequence}

\item{start_loc}{start codon location in peptide coordinates relative to whole transcript}

\item{stop_loc}{stop codon location in peptide coordinates relative to whole transcript}
}
\value{
junction location in translated protein coordinates or tx_junc_loc if in UTR
}
\description{
junc_loc_pep
}
