% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mod_made}
\alias{mod_made}
\title{Finding an open-reading frame given a DNA character sequence and documenting the errors associated with attempting to do so.}
\usage{
mod_made(cds, cds_mod)
}
\arguments{
\item{cds}{reference cds as sorted dataframe}

\item{cds_mod}{junction-modified cds as sorted dataframe}
}
\value{
"changed" if the starts and ends are different, "normal" if the starts and ends are the same, "add_reg" if more elements in cds_mod than cds, sub_reg if less elements in cds_mod than cds
}
\description{
Finding an open-reading frame given a DNA character sequence and documenting the errors associated with attempting to do so.
}
