% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{modify_cds}
\alias{modify_cds}
\title{modifying the combined cds using the junction}
\usage{
modify_cds(combo_cds, target_junc, junc, start_exon, end_exon)
}
\arguments{
\item{junc}{the junction location information output from is_UTR}

\item{cds}{the combined cds to be modified}

\item{intron}{the intron coordinates used for modifying the cds}
}
\description{
modifying the combined cds using the junction
}
