% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{running_sum}
\alias{running_sum}
\title{running_sum}
\usage{
running_sum(vector)
}
\arguments{
\item{vector}{The vector of numbers}
}
\value{
the vector modified to reflect the incremented sum
}
\description{
running_sum
}
