% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{split_kmers}
\alias{split_kmers}
\title{split_kmers}
\usage{
split_kmers(kmers, save_dir)
}
\arguments{
\item{kmers}{The kmers character vector}

\item{save_dir}{The directory to save the kmer vector to.}
}
\value{
A vector detailing the number of kmers saved per file saved.
}
\description{
split_kmers
}
