% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{star_to_leaf}
\alias{star_to_leaf}
\title{star_to_leaf}
\usage{
star_to_leaf(STAR_file)
}
\arguments{
\item{STAR_file}{The star file to be converted}
}
\value{
converted junc_dat dataframe
}
\description{
star_to_leaf
}
