% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{which_codon}
\alias{which_codon}
\title{Finding an open-reading frame given a DNA character sequence and documenting the errors associated with attempting to do so.}
\usage{
which_codon(cod, tx_junc_loc, start_loc, stop_loc)
}
\arguments{
\item{cod}{DNAstring containing codon start and end coordinates}

\item{tx_junc_loc}{The junction location in the transcript}

\item{start_loc}{start codon location in peptide coordinates relative to whole transcript}

\item{stop_loc}{stop codon location in peptide coordinates relative to whole transcript}
}
\value{
junction location in translated protein coordinates or tx_junc_loc if in UTR
}
\description{
Finding an open-reading frame given a DNA character sequence and documenting the errors associated with attempting to do so.
}
