% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{adjust_width}
\alias{adjust_width}
\title{adjust_width}
\usage{
adjust_width(cds)
}
\arguments{
\item{cds}{A cds dataframe with vectors of numeric widths, ends, and starts}
}
\value{
The cds data with an adjusted width
}
\description{
adjust_width
}
