% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{choose_exons}
\alias{choose_exons}
\title{finding the exons that are associated with start and end of the junction}
\usage{
choose_exons(target_junc, exons_by_gene, gene_tar)
}
\arguments{
\item{exons_by_gene}{the exons per gene in the genome}

\item{target_junction}{the target junction to use for analysis}
}
\value{
the unique start and end junction exons
}
\description{
finding the exons that are associated with start and end of the junction
}
