% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{countKmers}
\alias{countKmers}
\title{returns the unique set of kmers from a vector of non-unique kmers with
integer counts representing counts of the kmer in each junction-tx pair
uses lapply to iterate over each seq in seqs_vec lapply will automatically
return a list with index in seqs_vec as list access index in returned list}
\usage{
countKmers(seqs_vec, K)
}
\arguments{
\item{seqs_vec}{vector of sequences (i.e. data_canon[,2])}

\item{K}{the kmer length}
}
\value{
a list of integer vectors with kmer character vectors as names
}
\description{
returns the unique set of kmers from a vector of non-unique kmers with
integer counts representing counts of the kmer in each junction-tx pair
uses lapply to iterate over each seq in seqs_vec lapply will automatically
return a list with index in seqs_vec as list access index in returned list
}
