% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getUniqKmers}
\alias{getUniqKmers}
\title{getUniqKmers}
\usage{
getUniqKmers(kmers)
}
\arguments{
\item{kmers}{character vector of non-unique kmers}
}
\value{
vector of character vectors representing unique list of kmers from the union of all sequences
}
\description{
getUniqKmers
}
