% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{split_kmers}
\alias{split_kmers}
\title{This script splits the kmer files into blocks of 2000000 or less and saves them to the out directory.
This helps with computation efficiency}
\usage{
split_kmers(kmers, save_dir)
}
\arguments{
\item{kmers}{The kmers character vector}

\item{save_dir}{The directory to save the kmer vector to.}
}
\value{
A vector detailing the number of kmers saved per file saved.
}
\description{
This script splits the kmer files into blocks of 2000000 or less and saves them to the out directory.
This helps with computation efficiency
}
