% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{which_junctions_overlap}
\alias{which_junctions_overlap}
\title{which_junctions_overlap}
\usage{
which_junctions_overlap(curr_introns)
}
\arguments{
\item{curr_introns}{a list of junctions corresponding to the current gene being analyzed}
}
\value{
a list of non-overlapping junctions for the gene with row indeces dictating the associated row index of the curr_introns list
}
\description{
which_junctions_overlap
}
